relationships with foreign key, securing application with authentication

+++ taskmate/toDoList/views.py
@@ -23,13 +23,26 @@ def toDoList(request):
     if request.method == "POST":
         form = ToDoForm(request.POST or None)
         if form.is_valid():
+            form.save(
+                commit=False
+            ).user_id = (
+                request.user
+            )  # this line delay the form form saving and accessing the user_id (user_id - this is the foreign key we created in model) to access the current user and after this i save my form below
+            # the commit=False attribute delay my save and then i access my user_id which i created in model earlier then make migrations here i use the foreign key field save the form with use of request.user
             form.save()
+            # the below also working as expected
+            # formInstance = form.save(commit=False)
+            # formInstance.user_id = request.user
+            # formInstance.save()
             messages.success(request, ("ToDo's added successfully!"))
         return redirect("toDoList")
     else:
-        toDos = (
-            TodoList.objects.all()
-        )  # if we use pagination then u must mention objects.all with paranthesis like this object.all() else we got error, if we dont use pagination then dont need to be mention paranthesis
+        # toDos = (
+        # TodoList.objects.all() # if i use objects.all() it will get all the datas from the ToDo model
+        # )  # if we use pagination then u must mention objects.all with paranthesis like this object.all() else we got error, if we dont use pagination then dont need to be mention paranthesis
+        toDos = TodoList.objects.filter(
+            user_id=request.user
+        )  # this filter takes the field as a parameter and the with filed and request value it will the records acccoring to it
         paginationIns = Paginator(
             toDos, 5
         )  # this paginator class accepts two arguments, 1 is which datas to paginate, 2 is page limit
@@ -64,20 +77,26 @@ def editToDo(request, todo_id):
 @login_required
 def completeToDo(request, todo_id):
     toDo = TodoList.objects.get(pk=todo_id)
-    if toDo.status == True:
-        toDo.status = False
+    if toDo.user_id == request.user:
+        if toDo.status == True:
+            toDo.status = False
+        else:
+            toDo.status = True
+        toDo.save()
+        messages.success(request, ("ToDo Marked as per your Wish!"))
     else:
-        toDo.status = True
-    toDo.save()
-    messages.success(request, ("ToDo Marked as per your Wish!"))
+        messages.error(request, ("Your are not allowed to access this Location!"))
     return redirect("toDoList")
 
 
 @login_required
 def deleteToDo(request, todo_id):
     selectToDo = TodoList.objects.get(pk=todo_id)
-    if selectToDo.delete():
-        messages.success(request, ("ToDo deleted Successfully!"))
+    if selectToDo.user_id == request.user:
+        if selectToDo.delete():
+            messages.success(request, ("ToDo deleted Successfully!"))
+    else:
+        messages.error(request, ("Your are not allowed to access this Location!"))
         return redirect("toDoList")
